Unathi will no longer explode their groin for unimportant reasons

+++ code/game/objects/auras/regenerating_aura.dm
@@ -60,7 +60,7 @@
 	if(prob(grow_chance) && H.nutrition > grow_threshold)
 		for(var/limb_type in H.species.has_limbs)
 			var/obj/item/organ/external/E = H.organs_by_name[limb_type]
-			if(E && E.organ_tag != BP_HEAD && !E.vital && !E.is_usable())	//Skips heads and vital bits...
+			if(E && E.organ_tag != BP_HEAD && !(E.organ_tag == BP_GROIN && !E.is_stump()) && !E.vital && !E.is_usable())	//Skips heads and vital bits...
 				E.removed()			//...because no one wants their head to explode to make way for a new one.
 				qdel(E)
 				E= null
